requirements.txt
jishaku
discord.py==2.3.1
python-dotenv